requirements.txt
pyro == 1.8.4
pytorch == 2.0.0
pandas == 2.0.1
numpy == 1.24.3
matplotlib == 3.7.1
seaborn == 0.12.2
graphviz == 0.20.1
